+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @foundation-ui/hooks
 
+## 1.2.1
+
+### Patch Changes
+
+- Dependencies updates, Update window definition in ColorMode, remove obselete provider
+
 ## 1.2.0
 
 ### Minor Changes