+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @foundation-ui/styles
 
+## 1.3.1
+
+### Patch Changes
+
+- Dependencies updates, Update window definition in ColorMode, remove obselete provider
+
 ## 1.3.0
 
 ### Minor Changes